requirements.txt
aiodns
aiohttp
aiohttp-cors
alabaster
async-timeout
Babel
behave
cchardet
cffi
chardet
cryptography
docutils
idna
imagesize
Jinja2
MarkupSafe
multidict
parse
parse-type
pkg-resources
pyasn1
pycares
pycparser
Pygments
PyJWT
pytz
PyYAML
requests
six
snowballstemmer
Sphinx
uvloop
virtualenv
yarl
